+++ Kbuild
@@ -42,28 +42,28 @@ endif
 
 obj-$(CONFIG_LUNATIK_RUN) += lunatik_run.o
 
-obj-$(CONFIG_LUNATIK_DEVICE) += lib/luadevice.o
-obj-$(CONFIG_LUNATIK_LINUX) += lib/lualinux.o
-obj-$(CONFIG_LUNATIK_NOTIFIER) += lib/luanotifier.o
-obj-$(CONFIG_LUNATIK_SOCKET) += lib/luasocket.o
-obj-$(CONFIG_LUNATIK_RCU) += lib/luarcu.o
-obj-$(CONFIG_LUNATIK_THREAD) += lib/luathread.o
-obj-$(CONFIG_LUNATIK_FIB) += lib/luafib.o
-obj-$(CONFIG_LUNATIK_DATA) += lib/luadata.o
-obj-$(CONFIG_LUNATIK_PROBE) += lib/luaprobe.o
-obj-$(CONFIG_LUNATIK_SYSCALL) += lib/luasyscall.o
-obj-$(CONFIG_LUNATIK_XDP) += lib/luaxdp.o
-obj-$(CONFIG_LUNATIK_FIFO) += lib/luafifo.o
-obj-$(CONFIG_LUNATIK_XTABLE) += lib/luaxtable.o
-obj-$(CONFIG_LUNATIK_NETFILTER) += lib/luanetfilter.o
-obj-$(CONFIG_LUNATIK_COMPLETION) += lib/luacompletion.o
-obj-$(CONFIG_LUNATIK_CRYPTO_SHASH) += lib/luacrypto_shash.o
-obj-$(CONFIG_LUNATIK_CRYPTO_SKCIPHER) += lib/luacrypto_skcipher.o
-obj-$(CONFIG_LUNATIK_CRYPTO_AEAD) += lib/luacrypto_aead.o
-obj-$(CONFIG_LUNATIK_CRYPTO_RNG) += lib/luacrypto_rng.o
-obj-$(CONFIG_LUNATIK_CRYPTO_COMP) += lib/luacrypto_comp.o
-obj-$(CONFIG_LUNATIK_CPU) += lib/luacpu.o
-obj-$(CONFIG_LUNATIK_HID) += lib/luahid.o
-obj-$(CONFIG_LUNATIK_SIGNAL) += lib/luasignal.o
-obj-$(CONFIG_LUNATIK_BYTEORDER) += lib/luabyteorder.o
+obj-$(CONFIG_LUNATIK_DEVICE) += lib/lunadevice.o
+obj-$(CONFIG_LUNATIK_LINUX) += lib/lunalinux.o
+obj-$(CONFIG_LUNATIK_NOTIFIER) += lib/lunanotifier.o
+obj-$(CONFIG_LUNATIK_SOCKET) += lib/lunasocket.o
+obj-$(CONFIG_LUNATIK_RCU) += lib/lunarcu.o
+obj-$(CONFIG_LUNATIK_THREAD) += lib/lunathread.o
+obj-$(CONFIG_LUNATIK_FIB) += lib/lunafib.o
+obj-$(CONFIG_LUNATIK_DATA) += lib/lunadata.o
+obj-$(CONFIG_LUNATIK_PROBE) += lib/lunaprobe.o
+obj-$(CONFIG_LUNATIK_SYSCALL) += lib/lunasyscall.o
+obj-$(CONFIG_LUNATIK_XDP) += lib/lunaxdp.o
+obj-$(CONFIG_LUNATIK_FIFO) += lib/lunafifo.o
+obj-$(CONFIG_LUNATIK_XTABLE) += lib/lunaxtable.o
+obj-$(CONFIG_LUNATIK_NETFILTER) += lib/lunanetfilter.o
+obj-$(CONFIG_LUNATIK_COMPLETION) += lib/lunacompletion.o
+obj-$(CONFIG_LUNATIK_CRYPTO_SHASH) += lib/lunacrypto_shash.o
+obj-$(CONFIG_LUNATIK_CRYPTO_SKCIPHER) += lib/lunacrypto_skcipher.o
+obj-$(CONFIG_LUNATIK_CRYPTO_AEAD) += lib/lunacrypto_aead.o
+obj-$(CONFIG_LUNATIK_CRYPTO_RNG) += lib/lunacrypto_rng.o
+obj-$(CONFIG_LUNATIK_CRYPTO_COMP) += lib/lunacrypto_comp.o
+obj-$(CONFIG_LUNATIK_CPU) += lib/lunacpu.o
+obj-$(CONFIG_LUNATIK_HID) += lib/lunahid.o
+obj-$(CONFIG_LUNATIK_SIGNAL) += lib/lunasignal.o
+obj-$(CONFIG_LUNATIK_BYTEORDER) += lib/lunabyteorder.o
