+++ lib/lunacompletion.c
