+++ lib/lunacpu.c
